third_party/blink/web_tests/platform/mac-mac11-arm64/external/wpt/webxr/layers/xrWebGLBinding_constructor.https-expected.txt

This is a testharness.js-based test.
FAIL Ensure that XRWebGLBinding's constructor throws appropriate errors using webgl assert_equals: Should get InvalidStateError for creating with inline session. expected "InvalidStateError" but got "TypeError"
FAIL Ensure that XRWebGLBinding's constructor throws appropriate errors using webgl2 assert_equals: Should get InvalidStateError for creating with inline session. expected "InvalidStateError" but got "TypeError"
Harness: the test ran to completion.
